+++ src/test/kotlin/io/github/pedrovvitor/talentintelligence/adapter/security/HmacCandidateSourceFingerprinterTest.kt
@@ -54,6 +54,6 @@ class HmacCandidateSourceFingerprinterTest {
 
     companion object {
         private val TENANT_A = TenantId(UUID.fromString("00000000-0000-0000-0000-000000000001"))
-        private val TENANT_B = TenantId(UUID.fromString("00000000-0000-0000-0000-000000000002"))
+        private val TENANT_B = TenantId(UUID.fromString("00000000-0000-0000-0000-000000000003"))
     }
 }